Refactor code for task manager feature and components
