+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/donljh/orbital22-gather#readme",
   "dependencies": {
     "concurrently": "^7.2.1",
-    "frontend": "file:client"
+    "frontend": "file:client",
+    "orbital22-gather": "file:server"
   }
 }
